# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation of the DEEL BackEnd Task EndPoints",
    "version": "1.0.6",
    "title": "Swagger DEEL BackEnd Task",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "william.mori@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/balances/deposit/{user_id}": {
      "post": {
        "tags": [
          "Transactions",
          "Profile"
        ],
        "summary": "Deposits money into the the the balance of a client, a client can't deposit more than 25% his total of jobs to pay. (at the deposit moment)",
        "description": " **Notes**\n\n**Note 1** The value to deposit is informed in cents. per example US$1 should be sent as 100\n\n**Note 2** : Client can't deposit more than 25% his total of jobs to pay. (at the deposit moment). If the client doesn't have values to pay. the restriction will be in the current balance and in the value to deposit. This rule is not applied for logged contractors\n\n**Note 3** : Besides the 25% rule, a user can only deposit if his balance >= the amount to deposit\n\n**Note 4** : The values come from the balance of the logged user to the balance of the informed user\n\n**Note 5** : The source user and destination user could be client or contractor, there is no restriction\n\n**Note 6** : The payment operation will be successful only if all the updates in the source user balance and the destination user balance were successful, if not the operation will fail and no data is updated\n\n",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "header",
            "description": "the profile id of the logged user",
            "required": true,
            "type": "integer"
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "ID of target profile to deposit",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "value",
            "required": true,
            "description": "The value to deposit in cents, per example US$1 is 100",
            "schema": {
              "type": "object",
              "required": [
                "value"
              ],
              "properties": {
                "value": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "400": {
            "description": "Returned  when:\n\n Invalid value to deposit or\n\n User has insufficient balance to deposit or\n\n Client trying to deposit more than 25% of his total of jobs to pay\n",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "401": {
            "description": "Informed profile_id doesn't match any profile",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "404": {
            "description": "Returned  when:\n\n Informed user to deposit was not found ",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "500": {
            "description": "Returned  when:\n\n An error occurs trying to update the Balance of source profile or\n\n An error occurs trying to update the Balance of target profile \n\n Note: No data is updated in these cases  ",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "enum": [
            "ok",
            "fail"
          ]
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  }
}